+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Hero from "./hero";
+import Hero from "../components/homepage/hero";
 
 export default function Home() {
   return (
